+++ src/main/java/com/birca/bircabackend/command/artist/presentation/ArtistController.java
@@ -3,8 +3,8 @@ package com.birca.bircabackend.command.artist.presentation;
 import com.birca.bircabackend.command.artist.application.ArtistService;
 import com.birca.bircabackend.command.artist.dto.FavoriteArtistRequest;
 import com.birca.bircabackend.command.artist.dto.InterestArtistRequest;
-import com.birca.bircabackend.command.auth.authentication.LoginMember;
-import com.birca.bircabackend.command.auth.authentication.RequiredLogin;
+import com.birca.bircabackend.command.auth.authorization.LoginMember;
+import com.birca.bircabackend.command.auth.authorization.RequiredLogin;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
