--- src/main/java/com/birca/bircabackend/command/auth/authentication/LoginMember.java
@@ -1,4 +1,0 @@
-package com.birca.bircabackend.command.auth.authentication;
-
-public record LoginMember(Long id) {
-}
