+++ src/main/java/com/birca/bircabackend/command/auth/authorization/LoginMember.java
@@ -1,0 +1,4 @@
+package com.birca.bircabackend.command.auth.authorization;
+
+public record LoginMember(Long id) {
+}
