+++ src/main/java/com/birca/bircabackend/query/controller/ArtistQueryController.java
@@ -1,6 +1,6 @@
 package com.birca.bircabackend.query.controller;
 
-import com.birca.bircabackend.command.auth.authentication.RequiredLogin;
+import com.birca.bircabackend.command.auth.authorization.RequiredLogin;
 import com.birca.bircabackend.query.dto.ArtistGroupResponse;
 import com.birca.bircabackend.query.dto.ArtistResponse;
 import com.birca.bircabackend.query.dto.PagingParams;
