+++ src/test/java/com/birca/bircabackend/command/auth/RequiredLoginTestController.java
@@ -1,7 +1,7 @@
 package com.birca.bircabackend.command.auth;
 
-import com.birca.bircabackend.command.auth.authentication.LoginMember;
-import com.birca.bircabackend.command.auth.authentication.RequiredLogin;
+import com.birca.bircabackend.command.auth.authorization.LoginMember;
+import com.birca.bircabackend.command.auth.authorization.RequiredLogin;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
